requirements.txt.txt
django==1.9.1
djangorestframework
django-import-export